dev/backend/pyproject.toml
[project]
name = "backend"
version = "0.2.0"
description = "Add your description here"
authors = [
    { name = "nogi", email = "94681885+nogikun@users.noreply.github.com" }
]
dependencies = [
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.9",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "anyio==4.7.0",
    "attrs==24.2.0",
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "faiss-cpu==1.9.0.post1",
    "frozenlist==1.5.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.0",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "jiter==0.8.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.9",
    "langchain-community==0.3.9",
    "langchain-core==0.3.21",
    "langchain-openai>=0.2.12",
    "langchain-text-splitters==0.3.2",
    "langsmith==0.1.147",
    "marshmallow==3.23.1",
    "multidict==6.1.0",
    "mypy-extensions==1.0.0",
    "openai==1.57.0",
    "orjson==3.10.12",
    "packaging==24.2",
    "propcache==0.2.1",
    "pydantic>=2.10.3",
    "pydantic-settings==2.6.1",
    "pydantic-core==2.27.1",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.36",
    "tenacity==9.0.0",
    "tiktoken==0.8.0",
    "tqdm==4.67.1",
    "typing-inspect==0.9.0",
    "typing-extensions==4.12.2",
    "urllib3==2.2.3",
    "yarl==1.18.3",
    "numpy==1.26.4",
    "fastapi>=0.115.6",
    "uvicorn>=0.32.1",
    "firebase-admin>=6.6.0",
    "langfuse>=2.57.1",
    "langchain-qdrant>=0.2.0",
    "jsonpath-ng>=1.7.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]
